langchain/python/schema_extraction/pyproject.toml
[tool.poetry]
name = "prompt-tools-exp"
version = "0.1.0"
description = ""
authors = ["Daniel Chalef <131175+danielchalef@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.8"
langchain = "^0.0.242"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "^23.3.0"}
promptlayer = "^0.1.90"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
